requirements.txt
imagehash~=4.3.1
joker-cast~=0.5.2
joker-stream~=0.5.0
joker~=0.3.3
numpy~=2.2.1
pillow~=11.0.0
pymediainfo~=1.0
scipy~=1.14.1
volkanic~=0.5.3
